Revert "Add automatic request retry on CSRF failures. (#762)"
This reverts commit 5900f157fd7ca941f0487ee77b04ffd6e04d87eb.
