+++ app/static/js/controllers.js
@@ -1,4 +1,14 @@
-import { fetchWithCsrfRetry, getCsrfToken } from "./csrf.js";
+function getCsrfToken(doc = document) {
+  return getCsrfTokenElement(doc).getAttribute("content");
+}
+
+function getCsrfTokenElement(doc) {
+  return doc.querySelector("meta[name='csrf-token']");
+}
+
+function setCsrfToken(tokenValue) {
+  return getCsrfTokenElement(document).setAttribute("content", tokenValue);
+}
 
 class ControllerError extends Error {
   /**
@@ -53,6 +63,22 @@ async function processJsonResponse(response) {
   );
 }
 
+export async function refreshCsrfToken() {
+  return fetch("/")
+    .then(function (response) {
+      return response.text();
+    })
+    .then(function (html) {
+      const doc = new DOMParser().parseFromString(html, "text/html");
+      const csrfToken = getCsrfToken(doc);
+      setCsrfToken(csrfToken);
+      return Promise.resolve();
+    })
+    .catch(function (error) {
+      return Promise.reject("Failed to refresh CSRF token: " + error);
+    });
+}
+
 export async function getLatestRelease() {
   let route = "/api/latestRelease";
   return fetch(route, {
@@ -92,7 +118,7 @@ export async function shutdown(restart) {
   if (restart) {
     route = "/api/restart";
   }
-  return fetchWithCsrfRetry(route, {
+  return fetch(route, {
     method: "POST",
     headers: {
       "X-CSRFToken": getCsrfToken(),
@@ -122,7 +148,7 @@ export async function shutdown(restart) {
 
 export async function update() {
   let route = "/api/update";
-  return fetchWithCsrfRetry(route, {
+  return fetch(route, {
     method: "PUT",
     headers: {
       "X-CSRFToken": getCsrfToken(),
@@ -172,7 +198,7 @@ export async function determineHostname() {
 
 export async function changeHostname(newHostname) {
   const route = "/api/hostname";
-  return fetchWithCsrfRetry(route, {
+  return fetch(route, {
     method: "PUT",
     mode: "same-origin",
     cache: "no-cache",
@@ -251,7 +277,7 @@ export async function getVideoFps() {
 }
 
 export async function setVideoFps(videoFps) {
-  return fetchWithCsrfRetry("/api/settings/video/fps", {
+  return fetch("/api/settings/video/fps", {
     method: "PUT",
     mode: "same-origin",
     cache: "no-cache",
@@ -297,7 +323,7 @@ export async function getVideoJpegQuality() {
 }
 
 export async function setVideoJpegQuality(videoJpegQuality) {
-  return fetchWithCsrfRetry("/api/settings/video/jpeg_quality", {
+  return fetch("/api/settings/video/jpeg_quality", {
     method: "PUT",
     mode: "same-origin",
     cache: "no-cache",
@@ -327,7 +353,7 @@ export async function getDefaultVideoJpegQuality() {
 }
 
 export async function applyVideoSettings() {
-  return fetchWithCsrfRetry("/api/settings/video/apply", {
+  return fetch("/api/settings/video/apply", {
     method: "POST",
     mode: "same-origin",
     cache: "no-cache",
